+++ src/main/java/de/tum/in/www1/artemis/web/rest/ExamResource.java
@@ -21,6 +21,7 @@ import de.tum.in.www1.artemis.domain.User;
 import de.tum.in.www1.artemis.domain.exam.Exam;
 import de.tum.in.www1.artemis.domain.exam.ExerciseGroup;
 import de.tum.in.www1.artemis.domain.exam.StudentExam;
+import de.tum.in.www1.artemis.domain.participation.Participation;
 import de.tum.in.www1.artemis.repository.ExamRepository;
 import de.tum.in.www1.artemis.service.*;
 import de.tum.in.www1.artemis.service.dto.StudentDTO;
@@ -282,6 +283,31 @@ public class ExamResource {
 
         log.info("Generated {} student exams for exam {}", studentExams.size(), examId);
         return ResponseEntity.ok().body(studentExams);
+    }
+
+    /**
+     * POST /courses/{courseId}/exams/{examId}/studentExams/generate-participations : Generate the participation objects
+     * for all the student exams belonging to the exam
+     *
+     * @param courseId the course to which the exam belongs to
+     * @param examId   the exam to which the student exam belongs to
+     * @return ResponsEntity containing the list of generated participations
+     */
+    @PostMapping(value = "/courses/{courseId}/exams/{examId}/student-exams/generate-participations")
+    @PreAuthorize("hasAnyRole('INSTRUCTOR', 'ADMIN')")
+    public ResponseEntity<List<Participation>> generateParticipations(@PathVariable Long courseId, @PathVariable Long examId) {
+        log.info("REST request to generate participations for student exams of exam {}", examId);
+
+        // Checking if the user has necessary permissions to generate participations
+        Optional<ResponseEntity<List<Participation>>> courseAndExamAccessFailure = examAccessService.checkCourseAndExamAccess(courseId, examId);
+        if (courseAndExamAccessFailure.isPresent())
+            return courseAndExamAccessFailure.get();
+
+        List<Participation> generatedParticipations = examService.generateParticipations(examId);
+
+        log.info("Generated {} participations for student exams of exam {}", generatedParticipations.size(), examId);
+
+        return ResponseEntity.ok().body(generatedParticipations);
     }
 
     /**
