+++ src/main/java/de/tum/in/www1/artemis/web/rest/StudentExamResource.java
@@ -10,7 +10,6 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import de.tum.in.www1.artemis.domain.exam.StudentExam;
-import de.tum.in.www1.artemis.domain.participation.Participation;
 import de.tum.in.www1.artemis.service.ExamAccessService;
 import de.tum.in.www1.artemis.service.StudentExamAccessService;
 import de.tum.in.www1.artemis.service.StudentExamService;
@@ -82,30 +81,5 @@ public class StudentExamResource {
     public ResponseEntity<StudentExam> getStudentExamForConduction(@PathVariable Long courseId, @PathVariable Long examId, @PathVariable Long studentExamId) {
         log.debug("REST request to get student exam : {}", studentExamId);
         return studentExamAccessService.checkAndGetStudentExamAccessWithExercises(courseId, examId, studentExamId);
-    }
-
-    /**
-     * POST /courses/{courseId}/exams/{examId}/studentExams/generate-participations : Generate the participation objects
-     * for all the student exams belonging to the exam
-     *
-     * @param courseId the course to which the exam belongs to
-     * @param examId   the exam to which the student exam belongs to
-     * @return ResponsEntity containing the list of generated participations
-     */
-    @PostMapping(value = "/courses/{courseId}/exams/{examId}/studentExams/generate-participations")
-    @PreAuthorize("hasAnyRole('INSTRUCTOR', 'ADMIN')")
-    public ResponseEntity<List<Participation>> generateParticipations(@PathVariable Long courseId, @PathVariable Long examId) {
-        log.info("REST request to generate participations for student exams of exam {}", examId);
-
-        // Checking if the user has necessary permissions to generate participations
-        Optional<ResponseEntity<List<Participation>>> courseAndExamAccessFailure = examAccessService.checkCourseAndExamAccess(courseId, examId);
-        if (courseAndExamAccessFailure.isPresent())
-            return courseAndExamAccessFailure.get();
-
-        List<Participation> generatedParticipations = studentExamService.generateParticipations(examId);
-
-        log.info("Generated {} participations for student exams of exam {}", generatedParticipations.size(), examId);
-
-        return ResponseEntity.ok().body(generatedParticipations);
     }
 }
